Refactor package namme
