+++ .github/workflows/android.yml
@@ -33,7 +33,7 @@ jobs:
       run: chmod +x gradlew
       
     - name: Build with Gradle
-      run: ./gradlew build
+      run: ./gradlew assemble
 
     - name: Run tests
       run: ./gradlew test
