+++ app/src/main/AndroidManifest.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    package="com.thesohelshaikh.android_ci_test">
+    package="com.thesohelshaikh.androidcitest">
 
     <application
         android:allowBackup="true"
@@ -14,7 +14,7 @@
         android:theme="@style/Theme.Androidcitest"
         tools:targetApi="31">
         <activity
-            android:name=".MainActivity"
+            android:name="com.thesohelshaikh.androidcitest.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
