+++ app/src/main/java/com/thesohelshaikh/androidcitest/MainActivity.kt
@@ -1,7 +1,8 @@
-package com.thesohelshaikh.android_ci_test
+package com.thesohelshaikh.androidcitest
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
+import com.thesohelshaikh.android_ci_test.R
 
 class MainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
